langgraph-agentic-web-search/pyproject.toml
[project]
name = "langgraph-agentic-web-search"
version = "0.1.0"
description = ""
authors = [
    {name = "Jeff Taylor",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langgraph (>=0.2.70,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.4,<0.4.0)",
    "langchain-community (>=0.3.16,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
